Implementing windy world example.
Windy world is a world with a bridge, some water, and a goal. The agent
must reach the goal in order to receive the +100 reward. However, should
the agent drop in the water, it receives -100 reward. Also, -1 reward is
given for each step.

The agent must learn from the given rewards in order to learn the best
path towards the goal.