Creating client/server example using ZeroMQ.
ZeroMQ is an open-source project for IPC using sockets with steroids. It
allows easy creation of concurrent architectures using several
communication architectures such as client/server, subscriber/publisher,
and pipeline.
http://www.zeromq.org/

It needs to be installed in the OS with some package manager. For
instance, in Mac OS, `brew` quickly installs ZeroMQ.

```
brew install zeromq
```

Afterwards, the Python wrapper is available with `pip`.

```
sudo pip install pyzmq
```

This example implements a client/server structure that sends and
receives Python objects (using `pickle`). It should be used to
communicate the multiagent system with the Pacman simulator.