Bayesian inference example.
In this example, a 1,000 samples of a Gaussian distribution with mean =
3 and variance 10 are generated. This represents the measurements.

Next, each one of these measurements are feed to the grid. The
sensor uncertainty is represented as a Gaussian distribution centered
around the measurement and with variance of 30.

After feeding the entire data set, the infered probability distribution
is plotted. It is possible to see that the inferred Gaussian
distribution is centered around 3, the true mean.

Therefore, Bayesian inference is indeed occurring.