Creating map without ghosts.
As stated in #5, I created a map where Pacman is the only agent. It does
win games (obviously, since it cannot lose) but the behavior is
suboptimal. In 500 games, it scored about -2000 in approximatedly 5
games. It even scored -9000 once! Also, it does not appears to be
consistently learning, since its performance varies a lot.