Implement PacmanAdapterAgent using BaseAdapterAgent
This is the first step to create a unified interface that can be used by different RL agents in multiple environments. I managed to refactor the `PacmanAdapterAgent` to define its steps using the virtual methods required by `BaseAdapterAgent`.

Nothing was changed in the ghosts implementation so far.

As far as I could notice, the game is working without any problems.