Removing BehaviorCountMessage
Continuing with the work explained at #69, I'm removing the
`BehaviorCountMessage` since the controller shouldn't deal with logic
specific to one type of agent.

However, `PolicyMessage` will remain. This is a multiagent
*reinforcement learning*, hence policies are too common to be ignored.
However, I also implemented two methods in the `ControllerAgent`,
`get_policy` and `set_policy` with default (and dummy) implementations
to deal with policies.